examples/my-model/pyproject.toml
[project]
name = "my-model"
description = "A Disco simulation model."
version = "0.0.1"
dependencies = [
    "numpy~=2.3",
    "pandas~=2.3",
    "python-graphblas~=2025.2",
    "disco-toolbox>=0.0.2",
    "scipy>=1.17",
    "disco-tools>=0.1.8",
]

[build-system]
requires = [
    "setuptools>=69",
    "setuptools-scm>=8"
]
build-backend = "setuptools.build_meta"

[project.entry-points."disco.models"]
my-model = "my_model_pkg.plugin:model_package"
